app/backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.10
PyMuPDF>=1.24.11
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0
pydantic>=2.9.2
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.26.0
